feat(activities): Add an activity list to the vacation detail page
On the server side of things I have the vacation service also handling
getting the vacation activities as well. I also have the vacation
controller serving up the activities as well. IDK if it makes sense
to split it out at this point.

On the frontend I have the main vacation detail component handling
fetching for the activities. I think I should put that in its own
component. TBD if I change that.