feat(activity-media): Properly support uploading of activity media
In the previous commit we were able to "upload" activity media to the
server.

This was incomplete because it only stored the item in our remote
storage and it did not store anything in the database.

In this commit we properly hook up the remote storage with the database.
The database stores the URI of the item stored in the remote storage.
This refactor resulted in a new "ActivityService" that combines
the remote storage and database update logic into a single location.

The controller simply requests a save or retrieve from the activity
service and the service handles getting data from the database and
requesting/storing data in the remote storage.

There is currently a problem that I will solve later - I attempted
to implement "refresh" functionality for the activity media list
without refreshing the page. This works but there are errors in the
console about duplicate key entries which suggests that I did not
do this correctly. This will be my next change.